+++ scripts/macros/2014/classFeatures/illrigger/interdictBoons/hellishFrenzy.js
@@ -1,0 +1,36 @@
+import {itemUtils} from '../../../../../utils.js';
+async function added({trigger: {entity: item}}) {
+    await itemUtils.correctActivityItemConsumption(item, ['use'], 'balefulInterdict');
+}
+export let interdictBoonHellishFrenzy = {
+    name: 'Interdict Boons: Hellish Frenzy',
+    version: '1.3.71',
+    rules: 'legacy',
+    item: [
+        {
+            pass: 'created',
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
+        }
+    ],
+    config: [
+        {
+            value: 'playAnimation',
+            label: 'CHRISPREMADES.Config.PlayAnimation',
+            type: 'checkbox',
+            default: true,
+            category: 'animation'
+        }
+    ],
+    hasAnimation: true
+};